# src/routes/mobile/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string",
						"nullable": true
					},
					"isBusinessError": {
						"type": "boolean"
					},
					"errorDetail": {
						"type": "string",
						"nullable": true
					},
					"isEncrypted": {
						"anyOf": [
							{
								"type": "boolean"
							},
							{
								"type": "boolean",
								"enum": [
									false
								]
							}
						]
					},
					"resultApi": {},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"isBusinessError",
					"errorDetail",
					"resultApi"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteUserDto": {
				"properties": {
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordMobileDto": {
				"properties": {
					"oldPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"oldPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "sever-base-ctm",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/v1/mobile-app/users/{id}": {
			"get": {
				"operationId": "GetUserByIdPublic",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "Lấy thông tin người dùng công khai theo ID",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUserPermanent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "Xóa người dùng vĩnh viễn theo ID",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteUserDto"
							}
						}
					}
				}
			}
		},
		"/v1/mobile-app/users/change-password/{id}": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "Đổi mật khẩu cho người dùng theo ID",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordMobileDto"
							}
						}
					}
				}
			}
		},
		"/v1/public/categories": {
			"get": {
				"operationId": "GetAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/public/categories/{categoryId}": {
			"get": {
				"operationId": "GetCategoryById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/public/categories/parent/{parentId}": {
			"get": {
				"operationId": "GetCategoriesByParentId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "parentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}